MMAL: Indicate when picture came from still frame so advanced deinterlace can be disabled
MMAL Advanced deinterlace requires 3 frames of context so does not produce any output
from a dvd menu still.

We cannot easily submit the same frame multiple times (the MMAL buffer headers contain
linked list pointers), so the simple solution is to switch to the simpler
deinterlace that does not require context for stills.